+++ js/mootools/components/Accordion.js
@@ -21,9 +21,7 @@ Toolkit.Accordion = new Class({
         mode: 'click',
         defaultIndex: 0,
         multiple: false,
-        collapsible: false,
-        headerElement: '.accordion-header',
-        sectionElement: '.accordion-section'
+        collapsible: false
     },
 
     /**
@@ -38,8 +36,8 @@ Toolkit.Accordion = new Class({
         this.options = options = this.inheritOptions(this.options, element);
 
         // Fetch all the sections and headers
-        var sections = element.getElements(options.sectionElement),
-            headers = element.getElements(options.headerElement),
+        var sections = element.getElements('.' + vendor + 'accordion-section'),
+            headers = element.getElements('.' + vendor + 'accordion-header'),
             self = this;
 
         this.headers = headers;
@@ -71,13 +69,22 @@ Toolkit.Accordion = new Class({
                 .conceal();
         });
 
-        // Set events
-        this.events[options.mode + ' headers'] = 'onShow';
+        this.events = {
+            '{mode} element .@accordion-header': 'onShow'
+        };
 
         this.enable();
         this.fireEvent('init');
 
         this.jump(options.defaultIndex);
+    },
+
+    /**
+     * Reveal all sections before destroying.
+     */
+    doDestroy: function() {
+        this.headers.getParent().removeClass('is-active');
+        this.sections.removeProperty('style').reveal();
     },
 
     /**
@@ -178,9 +185,6 @@ Toolkit.Accordion = new Class({
 
 });
 
-/**
- * Defines a component that can be instantiated through accordion().
- */
 Toolkit.create('accordion', function(options) {
     return new Toolkit.Accordion(this, options);
 });